Add name="..." labels to introduction & naming.
These additions will allow folks a way to link a specific piece of the
documentation, for example if trying to send a link to a collegue to help
explain what a "modifier" is in BEM.

In the future, it may be good to provide a more user-friendly way to
find these hashtaggable links. In github for example, the header tags
in markdown are rendered with a "link" icon to their left when hovered,
that links to that header. This gives users something they can click on
to get the links, versus having to look through source.

None the less, I feel this commit is a good first start to enhance user's
ability to talk about & share information on BEM.